Redact sensitive query values, record the failed-over URL, and dedupe by instance
